admin/kafka_connections_service: Extract impl
No functional change.

Allow testing of the collectors.
